Remove superfluous type definition

+++ machinery-helper/machinery_helper_test.go
@@ -262,7 +262,7 @@ func TestAmendSize(t *testing.T) {
 	}
 	amendSize(&entry, 0)
 	want := int64(8240) // 2 directories (4096 bytes each) + 4 files (12 bytes each)
-	if int64(*entry.Size) != want {
+	if *entry.Size != want {
 		t.Errorf("entry.Size = '%v', want '%v'", *entry.Size, want)
 	}
 	want_files := 5
